moving the getIsPro check client side
