+++ hooks/useFileUpload.tsx
@@ -1,5 +1,6 @@
-import getIpfsJwt from '@/lib/actions/getIpfsJwt'
+import getIsPro from '@/lib/actions/getIsPro'
 import { MAX_FILE_SIZE, ONE_MB } from '@/lib/consts'
+import getIpfsJwt from '@/lib/getIpfsJwt'
 import { uploadFile } from '@/lib/ipfs/uploadFile'
 import isSupportedFileType from '@/lib/isSupportedFileType'
 import { useZoraCreateProvider } from '@/providers/ZoraCreateProvider'
@@ -22,9 +23,10 @@ const useFileUpload = () => {
     try {
       const file = event.target.files[0]
       if (!file) throw new Error()
-      const JWT = await getIpfsJwt(address)
+      const { isPro } = await getIsPro(address)
+      const JWT = isPro ? await getIpfsJwt() : null
 
-      if (!JWT && file.size > MAX_FILE_SIZE) {
+      if (!isPro && file.size > MAX_FILE_SIZE) {
         throw new Error(`File size exceeds the maximum limit of ${MAX_FILE_SIZE / ONE_MB}MB.`)
       }
       const mimeType = file.type
